+++ src/servlet/ModQAServlet.java
@@ -63,38 +63,44 @@ public class ModQAServlet extends HttpServlet {
 				String answer = request.getParameter("answer");
 
 				QasDAO qasDao = new QasDAO();
-				//Q&A登録
-				if (qasDao.insert(new QAs(0, question, answer))) {	// 登録成功
-					request.setAttribute("result",
-					new Result("登録成功！", "レコードを登録しました。", "/C3/ModQAServlet"));
+
+
+				if(request.getParameter("submit").equals("登録")) {
+					if (qasDao.insert(new QAs(0, question, answer))) {	// 登録成功
+						request.setAttribute("result",
+						new Result("登録成功！", "レコードを登録しました。", "/C3/ModQAServlet"));
+					}
+					else {												// 登録失敗
+						request.setAttribute("result",
+						new Result("登録失敗！", "レコードを登録できませんでした。", "/C3/ModQAServlet"));
+					}
 				}
-				else {												// 登録失敗
-					request.setAttribute("result",
-					new Result("登録失敗！", "レコードを登録できませんでした。", "/C3/ModQAServlet"));
+				else if (request.getParameter("submit").equals("更新")) {
+					// ここを改造する
+					if (qasDao.update(new QAs(qasId, question, answer))) {	// 登録成功
+						request.setAttribute("result",
+						new Result("更新成功！", "レコードを更新しました。", "/C3/ModQAServlet"));
+					}
+					else {												// 登録失敗
+						request.setAttribute("result",
+						new Result("更新失敗！", "レコードを更新できませんでした。", "/C3/ModQAServlet"));
+					}
+				}
+				else {
+
+					if(qasDao.delete(qasId)) {
+						request.setAttribute("result",
+								new Result("削除成功！", "レコードを削除しました。", "/C3/ModQAServlet"));
+					}
+					else {												// 登録失敗
+						request.setAttribute("result",
+						new Result("削除失敗！", "レコードを削除できませんでした。", "/C3/ModQAServlet"));
+					}
 				}
 
-				//Q&A更新
-				if (qasDao.update(new QAs(qasId, question, answer))) {	// 登録成功
-					request.setAttribute("result",
-					new Result("更新成功！", "レコードを更新しました。", "/C3/ModQAServlet"));
-				}
-				else {												// 登録失敗
-					request.setAttribute("result",
-					new Result("更新失敗！", "レコードを更新できませんでした。", "/C3/ModQAServlet"));
-				}
-
-				//Q&A削除
-				if(qasDao.delete(qasId)) {
-					request.setAttribute("result",
-							new Result("削除成功！", "レコードを削除しました。", "/C3/ModQAServlet"));
-				}
-				else {												// 登録失敗
-					request.setAttribute("result",
-					new Result("削除失敗！", "レコードを削除できませんでした。", "/C3/ModQAServlet"));
-				}
+				// 結果ページにフォワードする
 				RequestDispatcher dispatcher = request.getRequestDispatcher("/WEB-INF/jsp/result.jsp");
 				dispatcher.forward(request, response);
-
 
 	}
